Src/requirements.txt
dash==2.17.1
hx711==1.1.2.3
pandas==2.2.2
rpi-lgpio==0.6
tqdm==4.66.4
